More fixes for use cases
